Move sync_projects_from_gmc to sponsorship_compassion module
